Create merged BackwardReactionModel class.
